+++ App/_API/package.json
@@ -21,7 +21,6 @@
     "bcrypt": "^5.0.1",
     "dotenv": "^16.0.0",
     "express": "^4.17.2",
-    "express-flash": "^0.0.2",
     "express-session": "^1.17.2",
     "mysql2": "^2.3.3",
     "nodemailer": "^6.7.2",
