--- App/_API/routes/user.routes.js
@@ -1,12 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const userController = require('../controllers/user.controller');
-
-router.post('/getFoodsByUser', userController.postGetFoodsByUser);
-router.post('/getSportByUser', userController.postGetSportByUser);
-router.post('/getWaterByUser', userController.postGetWaterByUser);
-router.post('/postWaterByUser', userController.postWaterByUser);
-router.post('/postSportByUser', userController.postSportByUser);
-router.post('/postFoodByUser', userController.postFoodByUser);
-
-module.exports = router;
