+++ App/_API/server.js
@@ -2,7 +2,6 @@ const express = require('express');
 const app = express();
 const sequelize = require('sequelize');
 const bcrypt = require('bcrypt');
-const nodemailer = require('nodemailer');
 
 const PORT = 3001 || process.env.PORT;
 
@@ -12,7 +11,7 @@ app.use(express.json());
 app.use(require('./middleware/access.middleware').auth);
 
 app.use('/auth', require('./routes/auth.routes'));
-app.use('/', require("./routes/user.routes"));
+app.use('/', require('./routes/dashboard.routes'));
 
 app.listen(PORT, (err) => {
   if (err) {
